Pass dynamically the service price and service type to BookingCartBloc and fix service price type from int to double

+++ lib/modules/businessServiceReservation/models/service_reservation.dart
@@ -55,7 +55,7 @@ class Reservation {
   DateTime date;
   String customerTenant;
   String tenant;
-  int totalPrice;
+  double totalPrice;
   String createdBy;
   String updatedBy;
   String source;
